Fixed long error

+++ growbox/enum.py
@@ -104,7 +104,7 @@ class Enum(EnumBase):
         return int(self.value)
 
     def __long__(self):
-        return long(self.value)
+        return int(self.value)
 
     def __float__(self):
         return float(self.value)
